+++ packages/react/src/exampleTypes/Date.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Flex } from '../greyVest'
-import { contexturify } from '../utils/hoc'
+import { contexturifyWithoutLoader } from '../utils/hoc'
 import F from 'futil-js'
 import _ from 'lodash/fp'
 
@@ -195,4 +195,4 @@ let DateComponent = ({
   )
 }
 
-export default contexturify(DateComponent)
+export default contexturifyWithoutLoader(DateComponent)